% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the row signif_dat}
\usage{
eq_clean_data(dat = read_signif())
}
\arguments{
\item{dat}{A tibble of NOAA signif data with 47 variables, default for signif.txt}
}
\value{
A tibble containing 48 vaiables
}
\description{
A date column created by uniting the year, month, day and converting it to the Date class
    LATITUDE and LONGITUDE columns converted to numeric class
}
\details{
#refactor to standard evaluation instead of dplyr
}
\examples{
\dontrun{
clean_dat <- eq_clean_data()
}

}
