% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{function to create a HTML label for Location, Magnitude, and Total deaths}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{with significant earthquakes data}
}
\description{
This function should put together a character string for each earthquake
that will show the cleaned location (as cleaned by the eq_location_clean() function created in Module
 1), the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS), with boldface labels
 for each ("Location", "Total deaths", and "Magnitude").

 @details If an earthquake is missing values for any of these,
 both the label and the value should be skipped for that element of the tag.

 #refactor to stadard evauation without mutation
}
\examples{
\dontrun{
eq_location_clean() \%>\%
   dplyr::filter(COUNTRY \%in\% c('JAPAN')) \%>\%
   dplyr::filter(DATE > '2000-01-01') \%>\%
   dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
   eq_map(annot_col = "popup_text")
}


}
