% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{leaflet map for NOAA sifnificant earthquakes data}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{with significant earthquakes data}

\item{annot_col}{a column for annotation}
}
\description{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates
each point with in pop up window containing annotation data stored in a column of the data frame.
}
\details{
The user should be able to choose which column is used for the annotation
in the pop-up with a function argument named annot_col. Each earthquake should be
shown with a circle, and the radius of the circle should be proportional
to the earthquake's magnitude (EQ_PRIMARY).
}
\examples{
\dontrun{
eq_location_clean() \%>\%
   dplyr::filter(COUNTRY \%in\% c('JAPAN')) \%>\%
   dplyr::filter(DATE > '2000-01-01') \%>\%
   eq_map(annot_col="DATE")
}

}
