% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aethetics}

\item{data}{cleaned significant data}

\item{stat}{stat transformatiion}

\item{position}{adjustment}

\item{na.rm}{remove na}

\item{show.legend}{show.legend}

\item{inherit.aes}{override default}

\item{...}{...}
}
\value{
ggplot of a timeline object
}
\description{

}
\examples{

}
