% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map6.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\description{
new Geom for mapping a timeline for NOAA significant data
with a label to annotate top_n data points
}
\keyword{datasets}
