% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{read_signif}
\alias{read_signif}
\title{Read NOAA significant earthquke data}
\usage{
read_signif(path = system.file("extdata", "signif.txt", package =
  "MSDRCapstone"), url = FALSE)
}
\arguments{
\item{path}{defalut path set to inst/extdata/"signif.txt"}

\item{url}{set TRUE to download data from NOAA website}
}
\value{
a tibble containing 47 vaiables
}
\description{
Read NOAA significant earthquke data
}
\examples{
dat <- read_signif(url = TRUE)

}
